requirements.txt
requests==2.27.1
pandas==1.4.1
numpy==1.22.3
fastapi==0.75.0
farm-haystack==1.2.0
pdfplumber==0.6.0
Unidecode==1.3.4
beautifulsoup4==4.10.0
notebook==6.4.10
uvicorn
